requirements.txt
# ARES - Autonomous Resilient Enterprise Suite
# Python 3.12+ dependencies

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2

# AI & LLM
ollama==0.3.1
langchain==0.3.7
langchain-community==0.3.7
langgraph==0.2.40

# Vector Database & Search
chromadb==0.5.15
rank-bm25==0.2.2
sentence-transformers==3.3.1

# PII Detection & Privacy
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33
spacy==3.7.5

# Document Processing
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
markdown==3.7
beautifulsoup4==4.12.3

# Frontend
streamlit==1.40.0
streamlit-option-menu==0.3.13

# Utilities
loguru==0.7.2
python-multipart==0.0.12
aiofiles==24.1.0
httpx==0.27.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Development
black==24.8.0
ruff==0.6.4
mypy==1.11.1
